add area ,split, sample2 and match
